Added local-exec provisioner and bash script to fetch latest releases of rudder-server and rudder-transformer
fetch-latest-release.sh looks at the repo for the two projects and grabs the latest official release. Since terraform wants the un-tarred, un-gzipped rudder-server binary and a zip of rudder-tranformer, this grabs those release assets.

There's no option for cloning/building based on a SHA, branch name, etc. This is just to prevent needing to remember to update this repo each time a new release occurs on rudder-server and rudder-tranformer.